Prevent serving any files from outside the current working directory
At present, if you visit `http://localhost:7357//etc/passwd` you will see the file. It's very dangerous to run a web server as the current user with full access to the file system.

This change will return a 403 if any file outside of the current working directory is requested.

The user can turn this behavior off by changing the `unsafe_file_serving` config boolean to true.