IE6 compatibility: the stateChange handler is called with this === window rather than this === sendXHR
I have tested the other onreadystatechange handlers and they do not suffer from this issue. When using the htmlfile transport there seem to be a few different technologies being used.

The issue I had was that IE6 would connect and run tests but not report any results back to testem as it would buffer all of its requests as its first connect requests were never answered (it thought, as this.readyState === undefined).