Checking to make sure the server starts before continuing
If another process was already bound to port 7357, the built-in HTTP
server would fail to start up. This wasn't detected, so PhantomJS was
started anyway, and would try connecting to whatever process already
had 7357 open. And then everything would hang.

The error reporting here could probably use some improvement, but this
is enough to prevent it from hanging (which is brutal in CI mode).