Fix error messages for before hooks
The error mesages in the before_tests and onStart hooks relied on the
error object having stdout and stderr properties. Unfortunately they
don't - they use the Error object which only has a name and message
property. This meant that if one of those hooks failed the error would
say:

Error
undefined
undefined

I fixed this and removed a little duplication. There are holes here in
the testing, which I didn't fix.

See
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Error
for a description of the JS standard Error object.