Use postMessage for iframe-parent communication
Previous version accessed methods and data on `window.parent` directly,
but this is not safe. For instance, setting the `document.domain` within
a test will cause the browser to error whn the iframe later attemps to
access a property on `window.parent`.

This changes the code to use `window.postMessage` instead of accessing
`parent` directly. Because message-passing is asynchronous, a fair
amount of code is changing here to better handle that asynchrony.

The iframe will post the following messages to its parent window:
  * `get-id`: The parent will send the id back via another postMessage
  * `reload`: Called when restarting tests, reloads the parent window
  * `iframe-ready`: When the iframe is ready to process messages from
    the parent. Before it receives this event, the parent enqueues
    messages. On receiving 'iframe-ready' it will push all the enqueued
    messages to the iframe (which sends them on via socket.io to the
    api). After receiving this event, messages are no longer enqueued by
    the parent window, they are immediately sent to the iframe.
  * `no-connection-required`: If the id is -1, no socket connection is
    required and the parent need not send the iframe messages intended
    to be relayed over the socket

The iframe listens for these types of messages:
  * `emit-message`: When the parent window is passing it a message to
    send to the api over socket.io
  * `get-id`: When it receives the id from the parent

To be safe for browsers that do no support postMessage with non-strings as
messages, we use JSON.stringify and JSON.parse on either side of the
postMessage send/receive. Since the messages from the parent to iframe
may contain arbitrarily complex objects (due to `console.log` statements
in test code), we use `decycle` to break cyclical references before
sending the message. This is not necessary (and thus not used) when
posting from the iframe to the parent, because those types of messages and payloads
are not arbitrary (they all originate from the iframe's code).

all changes with logging

remove logging