Ensure failed tests are reported with XUnitReporter.
Prior to this, the `passed` flag for each test result was not actually
used to flag a given test as failed (we only checked `error`).

While reviewing the quasi-official implementation in the JUnit exporter
for `ant`, I noticed that error information should not be in `<failure>`
elements but they should be in `<error>` elements.

Implementation here:

* https://github.com/apache/ant/blob/master/src/main/org/apache/tools/ant/taskdefs/optional/junit/XMLJUnitResultFormatter.java
* https://github.com/apache/ant/blob/master/src/main/org/apache/tools/ant/taskdefs/optional/junit/XMLConstants.java