Revert change to use `<error>` instead of `<failure>` for XUnitReporter.
While I still maintain that the change is correct (and that it is useful
to know the difference between an assertion failure and an error), it is
a potentially breaking change (as reported in the commit comments).

This changes the element back while we figure out a better (compatible)
path forward.