Support running tests from a file: URL
There are a couple of changes here that enable (with some hoop-jumping) running tests from file: urls. My use case is electron -- testem is awesome, and I want to be able to test my electron apps using testem. More specifically, I'm doing some work on ember-electron, and ember is already set up to use testem. However, as you can see from the integration test, these changes will support electron in general, not just ember electron.

The primary changes are:
* Instead of using the URL '/testem/connection.html' to load the client into the iframe, parse out testem.js' script tag's src and construct a fully-qualified URL including the scheme, so that the URL from which testem.js is loaded determines where to find connection.html, not the URL from which the host HTML is loaded.
* Allow the host HTML to specify a global function to do custom discovery of the testem ID, so that it can be provided through some other mechanism than being the first path component.
* Provide some extra template variables to the launcher commands to make it easier to communicate with the testem server

With these two changes in place, a custom runner can use the new template variables to generate a URL pointing into the filesystem, munge the test HTML a bit to point to testem.js with a fully-qualified URL, stick the testem ID into a query param, and then run HTML with an extra <script> tag providing a global function to parse the testem ID out of the query param instead of looking in the path. Bing bang boom, electron runs all the code out of the filesystem, but reports everything seemlessly to testem as if it were an actual browser running code through the testem proxy server.