Fix multiple tab/reporter issue
When running in server mode, if something causes the browser to reload (e.g. the user changing the QUnit filters to run a subset of the tests, or a live reload in a framework like ember-cli), a new reporter and UI tab will be spawned. This was happening because the browser login code was only matching against still-pending runners.

We can make the check more resilient by instead of checking pending, just looking for a runner with the correct id that either doesn't have a socket, or whose socket is not connected. That allows the same browser to connect back to the same reporter any number of times across page refreshes.