Fix Mocha's duplicate failure logs
Previously, there was a bug in Mocha where the "fail" event would
not call the "test end" event, resulting in some errors not being
reported. This was accounted for on the Testem side by #1045, but
Mocha has since fixed this issue. This resulted in a bug where
test failures were being reported twice, so this commit fixes
the issue (#1129) by removing the original workaround.