Fix Mocha's duplicate error messages on failure
Mocha fires a "fail" event for test failures and hook failures, but
fires a "test end" event only for tests. Since we want to handle
hook failures as well, the "fail" event was accounted for in #1045.
But this caused error messages to be logged twice since the handler
code was also firing for the corresonding "test end" events that
are fired for every "fail" event. This commit removes the handler
code from the "test end" event for failed tests so that it logs
errors only once.